backend/pyproject.toml
[tool.poetry]
name = "stock-analysis-backend"
version = "0.1.0"
description = "股票分析平台后端 - FastAPI + MongoDB + Redis"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

# 禁用包模式，仅用于依赖管理
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.31.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
motor = "^3.3.0"
redis = {extras = ["hiredis"], version = "^5.0.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.1.0"
python-multipart = "^0.0.9"
email-validator = "^2.1.0"
python-dotenv = "^1.0.0"
apscheduler = "^3.10.0"
colorama = "^0.4.6"
pyyaml = "^6.0.0"
httpx-sse = "^0.4.0"
anyio = "^4.0.0"
# LangChain - 使用最新 v1.0+ 版本
langchain-core = "^1.0.0"
langchain = "^1.0.0"
langchain-openai = "^1.0.0"
langchain-anthropic = "^1.0.0"
langchain-mcp-adapters = "^0.1.0"
langchain-community = "^0.3.0"
# 市场数据源
tushare = "^1.2.60"
akshare = "^1.18.18"
pandas = "^2.0.0"
yfinance = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"
respx = "^0.21.0"
freezegun = "^1.4.0"
black = "^24.1.0"
ruff = "^0.1.0"
mypy = "^1.8.0"
requests = "^2.32.5"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=core/mcp",
    "--cov=modules/trading_agents/tools",
    "--cov-report=html",
    "--cov-report=term-missing",
]
markers = [
    "unit: 单元测试",
    "integration: 集成测试",
    "e2e: 端到端测试",
    "slow: 慢速测试",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

plugins = ["pydantic.mypy"]
